remove old argument stringifier
