--- core/stringification/argument-stringifier.ts
@@ -1,9 +1,0 @@
-import { Any, TypeMatcher } from "../spying";
-import { stringify } from "./stringify";
-
-export class ArgumentStringifier {
-
-    public stringify(argument: any) {
-        return stringify(argument);
-    }
-}
