+++ core/stringification/index.ts
@@ -1,7 +1,5 @@
-import { ArgumentStringifier } from "./argument-stringifier";
 import { stringify } from "./stringify";
 
 export {
-    ArgumentStringifier,
     stringify
 };
